% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\name{clonealign}
\alias{clonealign}
\title{Assign scRNA-seq to clone-of-origin}
\usage{
clonealign(gene_expression_data, copy_number_data, size_factors = NULL,
  max_iter = 100, rel_tol = 1e-05, gene_filter_threshold = 0,
  verbose = TRUE, bp_param = bpparam())
}
\arguments{
\item{gene_expression_data}{A matrix of gene counts or a
\code{SingleCellExperiment}. This should contain raw counts. See \code{details}.}

\item{copy_number_data}{A matrix or data frame of copy number calls for each clone.
See \code{details}.}

\item{size_factors}{Vector of cell size factors. If NULL computed using
\code{scran::computeSumFactors}}

\item{max_iter}{Maximum number of EM iterations before algorithm is terminated}

\item{rel_tol}{Relative tolerance (change in log-likelihood per EM iteration in percent) below which the
EM algorithm is considered converged}

\item{gene_filter_threshold}{Genes with mean counts below or equal to this threshold will
be filtered out (removes genes with no counts by default)}

\item{verbose}{Should warnings and EM convergence information be printed? Default TRUE}

\item{bp_param}{Parameters for parallel optimization of the Q function during EM. Default parameters
taken by call to \code{bpparam()}. See \code{?bpparam} for more details.}
}
\description{
\code{clonealign} assigns single cells (measured with RNA-seq) to their clones of origin, where
the clones have been inferred from ultra-shallow scDNA-seq and collated into copy number profiles.
}
\details{
\strong{Input format}

\code{gene_expression_data} must either be a \code{SingleCellExperiment} or \code{SummarizedExperiment}
with a \code{counts} assay
representing raw gene expression counts, or a cell by gene matrix of raw counts.

\code{copy_number_data} must either be a \code{matrix}, \code{data.frame} or \code{DataFrame} with a
row for each gene in \code{gene_expression_data} and a column for each of the clones.
If \code{colnames(copy_number_data)} is not \code{NULL} then these names will be used for each of
the clones in the final output.

\code{size_factors} should be a vector of size factors (one for each cell in \code{gene_expression_data}).
If \code{NULL} these are computed using \code{scran::computeSumFactors}.

\strong{EM convergence monitoring}

\strong{Multithreaded optimization}
}
\examples{
library(SingleCellExperiment)
data(example_sce)
copy_number_data <- rowData(example_sce)[,c("A", "B", "C")]
cal <- clonealign(example_sce, copy_number_data)
print(cal)
clones <- cal$clone
}
