% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\name{compute_ca_fit_mse}
\alias{compute_ca_fit_mse}
\title{Compute mean squared error of a clonealign fit}
\usage{
compute_ca_fit_mse(fit, Y, L, model_mu = TRUE, random_clones = FALSE)
}
\value{
The mean square error of the clonealign fit on the given expression data using
the provided clones
}
\description{
Compute mean squared error of a clonealign fit
}
\keyword{internal}
