% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\name{compute_correlations}
\alias{compute_correlations}
\title{Post-hoc compute correlations between copy number and expression}
\usage{
compute_correlations(Y, L, clones)
}
\arguments{
\item{Y}{Cell by gene count matrix}

\item{L}{Gene by clone copy number matrix}

\item{clones}{Assigned clones}
}
\value{
A vector of correlations between the copy number and assigned gene expression
}
\description{
Post-hoc compute correlations between copy number and expression
}
\keyword{internal}
