% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-specific.R
\name{construct_ai_likelihood}
\alias{construct_ai_likelihood}
\title{Construct allele specific complete data likelihood}
\usage{
construct_ai_likelihood(clone_allele_ph, alt_ph, cov_ph, dtype, C, N)
}
\description{
Returns cell-by-clone complete likelihood
}
