% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\docType{data}
\name{df_cnv}
\alias{df_cnv}
\title{Example raw copy number data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 435 rows and 5 columns.}
\usage{
df_cnv
}
\description{
An example data frame with copy number calls by
region for three clones (A, B, C)
}
\examples{
data(df_cnv)
}
\keyword{datasets}
