% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\name{evaluate_clonealign}
\alias{evaluate_clonealign}
\title{Evaluate the quality of a clonealign fit}
\usage{
evaluate_clonealign(gene_expression_data, copy_number_data, clonealign_fit,
  prop_holdout = 0.2, n_samples = 2, s = NULL, ...)
}
\arguments{
\item{gene_expression_data}{Either a \code{SingleCellExperiment} or matrix of counts, same as input to \code{clonealign}}

\item{copy_number_data}{A gene-by-clone matrix of copy numbers, the same as the input to \code{clonealign}}

\item{clonealign_fit}{The fit returned by a call to \code{clonealign()} on the full geneset}

\item{prop_holdout}{The proportion of genes to hold out as a \emph{test} set for predicting gene expression}

\item{n_samples}{The number of random permutations to establish a null distribution of predictive performance}

\item{s}{Vector of cell size factors - defaults to the total counts per cell}

\item{...}{Additional arguments to pass to the \code{clonealign} call}
}
\value{
A list object describing the evaluation results. See \code{details}
}
\description{
Use mean squared error of predicting the expression of held-out genes to evaluate the quality of a clonealign fit.
}
\details{
This evaluation function is built around the idea of how good predicted expression is under the model given
the inferred (assigned) clones compared to how well you could predict expression given a random clonal assignment.

\strong{Evaluations performed}

\enumerate{
\item On the \emph{full} dataset, the mean square error is compared to the randomized error, and a message
printed about the ratio of the two errors and the proportion of time the observed error was less than the
error under a null distribution. If either the error under the null is smaller than the observed, or is
smaller some percentage of time, then the fit has real problems and should be abandoned as it suggests the
model is stuck in a local maxima (which could happen if the proposed clones aren't actually present in the
RNA-seq).
\item A certain proportion of genes (as decided by the \code{prop_holdout} parameter) are held out as a \emph{test}
set, and the clonealign fit is re-performed on the \code{1 - prop_holdout} proportion of genes. The function will then
print an agreement table of clone assignments between the full table and the reduced table. If these vastly disagree
for only a small change in gene set (ie \code{prop_holdout} is around 0.1 or 0.2), then the fit may be unreliable
as it is sensitive to the genes inputted.
\item The same metrics from (1) are then printed where the evaluations are performed on the heldout set. Again,
if the observed mean squared error given the clonealign fit isn't less than the null mean squared error
a large proportion of the time, the fit may be unreliable.

}

\strong{Object returned}

Everything computed above is returned in a list object with the following entries:

\itemize{
\item full_clonealign_fit - the original clonealign fit on the full gene set that was passed in as the
\code{clonealign_fit} argument
\item full_observed_mse - the observed mean square error using the full gene set
\item full_null_mse - A vector of mean square error under the randomized (permuted) distribution
\item reduced_clonealign_fit - a clonealign fit on only the reduced (train) set of genes
\item heldout_genes - the names of the genes held out (test set) for evaluating predictive performance
out-of-sample
\item kept_names - the names of retained genes as part of the reduced (train) set
\item heldout_observed_mse - the observed mean square error on the heldout (test) set of genes
\item heldout_null_mse - a vector of mean square errors under a null distribution of randomly permuting the clones

}
}
\examples{
library(SingleCellExperiment)
data(example_clonealign_fit)
data(example_sce)
copy_number_data <- rowData(example_sce)[,c("A", "B", "C")]
evaluate_clonealign(example_sce, copy_number_data, example_clonealign_fit)

}
