% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\docType{data}
\name{example_clonealign_fit}
\alias{example_clonealign_fit}
\title{Example clonealign_fit}
\format{An object of class \code{clonealign_fit} of length 2.}
\usage{
example_clonealign_fit
}
\description{
An example \code{clonealign_fit} using the data in \code{example_sce}.
}
\examples{
data(example_clonealign_fit)
}
\seealso{
example_sce
}
\keyword{datasets}
