% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\docType{data}
\name{example_sce}
\alias{example_sce}
\title{Example SingleCellExperiment}
\format{An object of class \code{SingleCellExperiment} with 100 rows and 200 columns.}
\usage{
example_sce
}
\description{
An example \code{SingleCellExperiment} for 100 genes and 200 cells.
Copy number data is in \code{rowData(example_sce)[, c("A", "B", "C",]}
}
\examples{
data(example_sce)
}
\seealso{
example_clonealign_fit
}
\keyword{datasets}
