% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\name{get_outlying_genes}
\alias{get_outlying_genes}
\title{Get outlier genes}
\usage{
get_outlying_genes(Y, nmads)
}
\arguments{
\item{Y}{Cell by gene matrix of counts}

\item{nmads}{Number of mads above which gene is considered outlier}
}
\description{
Get outlier genes
}
\keyword{internal}
