% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-em.R
\name{inference_em}
\alias{inference_em}
\title{Expectation-maximization for assigning scRNA-seq data
to clone-of-origin}
\usage{
inference_em(Y, L, s = NULL, max_iter = 100, rel_tol = 1e-05,
  gene_filter_threshold = 0, verbose = TRUE, multithread = TRUE,
  bp_param = bpparam())
}
\arguments{
\item{Y}{cell-by-gene expression matrix of raw counts}

\item{L}{gene-by-clone matrix of copy number variation}

\item{s}{Vector of cell size factors. If NULL computed using
\code{scran::computeSumFactors}}

\item{max_iter}{Maximum number of EM iterations before algorithm is terminated}

\item{rel_tol}{Relative tolerance in percent below which the log-likelihood is considered converged}

\item{gene_filter_threshold}{Genes with mean counts below or equal to this threshold will
be filtered out (removes genes with no counts by default)}

\item{verbose}{Logical - should convergence information be printed?}

\item{multithread}{Should the M-step be performed in parallel using \code{BiocParallel}? Default \code{TRUE}}

\item{bp_param}{Parameters for multithreaded optimization of Q function. See \code{?bpparam()}}
}
\value{
A list with ML estimates for each of the model parameters
}
\description{
Expectation-maximization for assigning scRNA-seq data
to clone-of-origin
}
\keyword{internal}
