% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-tflow.R
\name{inference_tflow}
\alias{inference_tflow}
\title{EM inference with tensorflow}
\usage{
inference_tflow(Y_dat, L_dat, max_iter = 100, rel_tol = 1e-05,
  learning_rate = 0.1, gene_filter_threshold = 0, x = NULL,
  clone_allele = NULL, cov = NULL, ref = NULL, fix_alpha = FALSE,
  dtype = c("float32", "float64"), saturate = TRUE,
  saturation_threshold = 6, K = 1, mc_samples = 1, verbose = TRUE,
  initial_shrink = 5, seed = NULL, data_init_mu = data_init_mu)
}
\arguments{
\item{Y_dat}{Cell by gene matrix of counts}

\item{L_dat}{Gene by clone matrix of copy number}
}
\value{
A list of maximum likelihood parameter estimates and log likelihood values
}
\description{
EM inference with tensorflow
}
\keyword{internal}
