% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\name{plot_mean_dispersion}
\alias{plot_mean_dispersion}
\title{Plot mean dispersion relationship}
\usage{
plot_mean_dispersion(em)
}
\arguments{
\item{em}{An object returned by a call to \code{clonealign}}
}
\value{
A \code{ggplot2} plot showing the dispersion as a function of mean
}
\description{
Plot mean dispersion relationship
}
\examples{
data(example_clonealign_fit)
plot_mean_dispersion(example_clonealign_fit)
}
