% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_for_clonealign}
\alias{preprocess_for_clonealign}
\title{Preprocess data for input to clonealign}
\usage{
preprocess_for_clonealign(
  gene_expression_data,
  copy_number_data,
  min_counts_per_gene = 20,
  min_counts_per_cell = 100,
  remove_outlying_genes = TRUE,
  nmads = 10,
  max_copy_number = 6,
  remove_genes_same_copy_number = TRUE
)
}
\arguments{
\item{gene_expression_data}{Input gene expression data. See \code{clonealign} for details}

\item{copy_number_data}{Input copy number data. See \code{clonealign} for details}

\item{min_counts_per_gene}{Minimum counts per gene for the gene to pass filtering}

\item{min_counts_per_cell}{Minimum counts per cell for the cell to pass filtering}

\item{remove_outlying_genes}{Logical - should genes whose expression is an outlier wrt
all others be removed?}

\item{nmads}{The number of median absolute deviations (MADs) the per-gene mean of the 
raw counts is from the overall mean to be considered an outlier}

\item{max_copy_number}{Maximum copy number per gene to retain (see "saturation" under original paper)}

\item{remove_genes_same_copy_number}{Logical - should genes with the same copy number in all clones be removed?}
}
\value{
A list with entries \code{gene_expression_data} and \code{copy_number_data}
for input to clonealign, along with names of the retained genes and cells after filtering.
}
\description{
Preprocess data for input to clonealign, filtering for cells and genes with minimum counts, along
with removing outlying genes, and genes with the same copy number between clones.
}
\examples{
library(SummarizedExperiment)
data("example_sce")
L <- rowData(example_sce)[,c("A", "B", "C")]
ca_data <- preprocess_for_clonealign(example_sce, L)
}
