% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{process_cnv_data}
\alias{process_cnv_data}
\title{Preprocess copy number data}
\usage{
process_cnv_data(copy_number_data)
}
\value{
The processed copy number data
}
\description{
Preprocess copy number data
}
\keyword{internal}
