% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{process_gene_expression_data}
\alias{process_gene_expression_data}
\title{Check and pre-process gene expression data}
\usage{
process_gene_expression_data(gene_expression_data)
}
\value{
The processed gene expression data
}
\description{
Check and pre-process gene expression data
}
\keyword{internal}
