% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-em.R
\name{Q_g}
\alias{Q_g}
\title{Computes Q(theta|theta^(t))
(function to be optimised under EM)}
\usage{
Q_g(pars, y, l, gamma, data)
}
\arguments{
\item{pars}{Parameters to optimise}

\item{y}{Gene expression for gene}

\item{l}{Copy number profiles for gene}

\item{gamma}{Expectation of clone assignments at current EM step}

\item{data}{Data used}
}
\value{
The g'th term in the expected complete data log likelihood
}
\description{
Computes Q(theta|theta^(t))
(function to be optimised under EM)
}
\keyword{internal}
