% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-em.R
\name{Qgr_g}
\alias{Qgr_g}
\title{Gradient of Q(theta|theta^(t)) (function to be optimised under EM)}
\usage{
Qgr_g(pars, y, l, gamma, data)
}
\arguments{
\item{pars}{Parameters to optimise}

\item{y}{Gene expression for gene}

\item{l}{Copy number profiles for gene}

\item{gamma}{Expectation of clone assignments at current EM step}

\item{data}{Data used}
}
\value{
The gradient g'th term in the expected complete data log likelihood
}
\description{
Gradient of Q(theta|theta^(t)) (function to be optimised under EM)
}
\keyword{internal}
