% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-tflow.R
\name{recompute_clone_assignment}
\alias{recompute_clone_assignment}
\title{Re-computes map clone assignment at a given threshold}
\usage{
recompute_clone_assignment(ca, clone_assignment_probability = 0.95)
}
\arguments{
\item{ca}{A clonealign fit}
}
\value{
A clonealign fit with clones re-called at different probabilities
}
\description{
Re-computes map clone assignment at a given threshold
}
\keyword{internal}
