% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\name{run_clonealign}
\alias{run_clonealign}
\title{Run clonealign across a range of initializations}
\usage{
run_clonealign(
  gene_expression_data,
  copy_number_data,
  initial_shrinks = c(0, 5, 10),
  n_repeats = 3,
  print_elbos = TRUE,
  ...
)
}
\arguments{
\item{gene_expression_data}{A matrix of gene counts or a
\code{SingleCellExperiment}. See \code{?clonealign}}

\item{copy_number_data}{A matrix or data frame of copy number calls for each clone.
See \code{?clonealign}.}

\item{initial_shrinks}{Initial shrinkages for the clone assignment variational parameters}

\item{n_repeats}{Number of fits to perform at each initial shrink}

\item{print_elbos}{Logical - should the ELBOs inferred be printed?}

\item{...}{Additional arguments to pass to \code{clonealign(...)}}
}
\value{
The \code{clonealign_fit} object for the fit that maximizes the ELBO. 
See \code{?clonealign} for details.
}
\description{
Run \code{clonealign} across a range of initializations and select the fit 
that acheives the best evidence lower bound (ELBO).
}
\details{
This function essentially wraps \code{clonealign} and can be interacted with as such. The
parameter \code{initial_shrinks} controls how hard the variational parameters are initially
assigned, analagous to the E-step in EM. At 0, they are initialized evenly across clones,
while at 10 they are semi hard assigned to the most likely initial values.
}
\examples{
library(SummarizedExperiment)
data("example_sce")
L <- rowData(example_sce)[,c("A", "B", "C")]
run_clonealign(example_sce, L, initial_shrinks = c(1), n_repeats = 1)
}
