% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\name{saturate}
\alias{saturate}
\title{Saturate a copy number matrix above a certain threshold}
\usage{
saturate(x, threshold = 4)
}
\value{
The input clipped (or saturated) below threshold
}
\description{
Saturate a copy number matrix above a certain threshold
}
\keyword{internal}
