% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-em.R
\name{clone_assignment}
\alias{clone_assignment}
\title{Computes map clone assignment given EM object}
\usage{
clone_assignment(tflow_res)
}
\arguments{
\item{tflow_res}{List returned by \code{inference_tflow}}
}
\value{
A vector of maximum likelihood clone assignments
}
\description{
Computes map clone assignment given EM object
}
\keyword{internal}
