% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\name{clonealign}
\alias{clonealign}
\title{Assign scRNA-seq to clone-of-origin}
\usage{
clonealign(gene_expression_data, copy_number_data, s = NULL, max_iter = 100,
  rel_tol = 1e-05, gene_filter_threshold = 0, bp_param = bpparam())
}
\arguments{
\item{gene_expression_data}{A matrix of gene counts or a
\code{SingleCellExperiment}. See \link{details}.}

\item{copy_number_data}{A matrix of copy number calls for each clone or a \code{SummarizedExperiment}.
See \link{details}.}

\item{s}{Vector of cell size factors. If NULL computed using
\code{scran::computeSumFactors}}

\item{max_iter}{Maximum number of EM iterations before algorithm is terminated}

\item{rel_tol}{Relative tolerance (change in log-likelihood per EM iteration in percent) below which the 
EM algorithm is considered converged}

\item{gene_filter_threshold}{Genes with mean counts below or equal to this threshold will
be filtered out (removes genes with no counts by default)}

\item{bp_param}{Parameters for parallel optimization of the Q function during EM. Default parameters
taken by call to \code{bpparam()}. See \code{?bpparam} for more details.}
}
\description{
Assign scRNA-seq to clone-of-origin
}
\details{
\strong{Input format}
ABC
}
