% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonealign.R
\name{clonealign}
\alias{clonealign}
\title{Assign scRNA-seq to clone-of-origin}
\usage{
clonealign(gene_expression_data, copy_number_data, size_factors = NULL,
  max_iter = 100, rel_tol = 1e-05, gene_filter_threshold = 0,
  verbose = TRUE, bp_param = bpparam())
}
\arguments{
\item{gene_expression_data}{A matrix of gene counts or a
\code{SingleCellExperiment}. This should contain raw counts. See \code{details}.}

\item{copy_number_data}{A matrix or data frame of copy number calls for each clone.
See \code{details}.}

\item{size_factors}{Vector of cell size factors. If NULL computed using
\code{scran::computeSumFactors}}

\item{max_iter}{Maximum number of EM iterations before algorithm is terminated}

\item{rel_tol}{Relative tolerance (change in log-likelihood per EM iteration in percent) below which the
EM algorithm is considered converged}

\item{gene_filter_threshold}{Genes with mean counts below or equal to this threshold will
be filtered out (removes genes with no counts by default)}

\item{verbose}{Should warnings and EM convergence information be printed? Default TRUE}

\item{bp_param}{Parameters for parallel optimization of the Q function during EM. Default parameters
taken by call to \code{bpparam()}. See \code{?bpparam} for more details.}
}
\value{
An object of class \code{clonealign_fit}. The maximum likelihood estimates of the
clone assignment paramters are in the \code{clone} slot. Maximum likelihood estimates of
all model parameters are in the \code{ml_params} slot.
}
\description{
\code{clonealign} assigns single cells (measured with RNA-seq) to their clones of origin, where
the clones have been inferred from ultra-shallow scDNA-seq and collated into copy number profiles.
}
\details{
\strong{Input format}

\code{gene_expression_data} must either be a \code{SingleCellExperiment} or \code{SummarizedExperiment}
with a \code{counts} assay
representing raw gene expression counts, or a cell by gene matrix of raw counts.

\code{copy_number_data} must either be a \code{matrix}, \code{data.frame} or \code{DataFrame} with a
row for each gene in \code{gene_expression_data} and a column for each of the clones.
If \code{colnames(copy_number_data)} is not \code{NULL} then these names will be used for each of
the clones in the final output.

\code{size_factors} should be a vector of size factors (one for each cell in \code{gene_expression_data}).
If \code{NULL} these are computed using \code{scran::computeSumFactors}.

\strong{EM convergence monitoring}

Inference is performed using the EM algorithm (\url{https://en.wikipedia.org/wiki/Expectation-maximization_algorithm})
which uses the log marginal likelihood to monitor convergence. This is controlled using the
\code{rel_tol} parameter. When the difference between the new and old log marginal likelihoods normalized
by the absolute value of the old falls below \code{rel_tol}, the EM algorithm is considered converged.
The maximum number of iterations to acheive this is set using the \code{max_iter} parameter.

\strong{Multithreaded optimization}

The M-step of the EM algorithm consists of \eqn{G} different optimizations for \eqn{G} parameters. These optimizations
are performed in parallel using the \code{BiocParallel} package. The control over the parallel
computations are controlled using the \code{bp_param} argument. For fine control over this see
the \code{\link[BiocParallel]{bpparam}} function. By default \code{clonealign} uses the default parameters
that are returned by a call to \code{bpparam()}. To use only e.g. 1 core call \code{clonealign} with
\code{clonealign(..., bp_param = MulticoreParam(workers = 1))}.
}
\examples{
library(SingleCellExperiment)
data(example_sce)
copy_number_data <- rowData(example_sce)[,c("A", "B", "C")]
cal <- clonealign(example_sce, copy_number_data)
print(cal)
clones <- cal$clone
}
