% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-specific.R
\name{construct_ai_likelihood}
\alias{construct_ai_likelihood}
\title{Construct allele specific complete data likelihood}
\usage{
construct_ai_likelihood(clone_allele_ph, alt_ph, cov_ph, dtype, C, N)
}
\arguments{
\item{clone_allele_ph}{Tensorflow placeholder for copy number at each variant position}

\item{alt_ph}{Tensorflow placeholder for alternative count}

\item{cov_ph}{Tensorflow placeholder for coverage count}

\item{dtype}{Tensorflow dtype}

\item{C}{Number of clones}

\item{N}{Number of cells}
}
\description{
Returns cell-by-clone complete likelihood
}
\keyword{internal}
