% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_outlying_genes}
\alias{get_outlying_genes}
\title{Get outlier genes}
\usage{
get_outlying_genes(Y, nmads)
}
\arguments{
\item{Y}{Cell by gene matrix of counts}

\item{nmads}{Number of mads above which gene is considered outlier}
}
\value{
A logical vector whether each gene (as represented by a column of Y) 
is an outlier
}
\description{
Get outlier genes
}
\examples{
Y <- matrix(rpois(200, 10), 20, 10)
clonealign:::get_outlying_genes(Y, 3)
}
\keyword{internal}
