% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-tflow.R
\name{inference_tflow}
\alias{inference_tflow}
\title{EM inference with tensorflow}
\usage{
inference_tflow(Y_dat, L_dat, max_em_iter = 50, rel_em_tol = 1e-05,
  max_adam_iter = 200, rel_adam_tol = 1e-06, gene_filter_threshold = 0,
  verbose = TRUE)
}
\description{
EM inference with tensorflow
}
