% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-em.R
\name{inference_tflow}
\alias{inference_tflow}
\title{EM inference with tensorflow}
\usage{
inference_tflow(Y_dat, L_dat, max_iter = 100, rel_tol = 1e-05,
  gene_filter_threshold = 0, verbose = TRUE, multithread = TRUE,
  bp_param = bpparam())
}
\description{
EM inference with tensorflow
}
