% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-em.R
\name{p_pi}
\alias{p_pi}
\title{Computes gamma_{nc} = p(pi_n = c), returning
N by C matrix}
\usage{
p_pi(data, params)
}
\arguments{
\item{data}{Input data}

\item{params}{Model parameters}
}
\value{
The probability that each cell belongs to each clone, as a matrix
}
\description{
Computes gamma_{nc} = p(pi_n = c), returning
N by C matrix
}
\keyword{internal}
