% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-em.R
\name{Q_g}
\alias{Q_g}
\title{Computes Q(theta|theta^(t))
(function to be optimised under EM)}
\usage{
Q_g(pars, y, l, gamma, data)
}
\description{
Computes Q(theta|theta^(t))
(function to be optimised under EM)
}
